+++ sycamore/__init__.py
@@ -1,3 +1,3 @@
 """Sycamore package metadata."""
 
-__version__ = "0.11.0"
+__version__ = "0.11.1"